docs: Increase verbosity of shell example
When reading through the shell example for manipulating torrents, it's
assumed that most users will either copy/paste the command or analyze it
to understand the mechanism of operation.

To make it easier to cross reference the command against `rtorrent
--help`, this commit converts the arguments to their longer form which
will allow users to dissect them one by one.

This should have no affect on users simply copying and pasting.